LayoutTests/inspector/dom-debugger/xhr-breakpoints-expected.txt

Tests for XHR breakpoints.


== Running test suite: XHRBreakpoints
-- Running test case: BreakOnAllXMLHttpRequests
Set pause on all XMLHttpRequests.
Sending XMLHttpRequest.
PAUSED:
PASS: Pause reason should be XHR.
PASS: Pause data URL should be empty.
CALL STACK:
0: [F] loadResourceXHR
1: [P] Global Code
-- Running test teardown.

-- Running test case: BreakOnXMLHttpRequestContainingText
PASS: Added breakpoint for URL: data
PASS: Breakpoint should not be disabled initially.
Sending XMLHttpRequest.
PAUSED:
PASS: Pause reason should be XHR.
PASS: Pause data URL should match breakpoint URL.
CALL STACK:
0: [F] loadResourceXHR
1: [P] Global Code
-- Running test teardown.

-- Running test case: BreakOnXMLHttpRequestMatchingRegularExpression
PASS: Added breakpoint for URL: data[A-Z]*.(json|txt|png)
PASS: Breakpoint should not be disabled initially.
Sending XMLHttpRequest.
PAUSED:
PASS: Pause reason should be XHR.
PASS: Pause data URL should match breakpoint URL.
CALL STACK:
0: [F] loadResourceXHR
1: [P] Global Code
-- Running test teardown.

-- Running test case: ShouldNotPauseOnDisabledBreakpoint
PASS: Added breakpoint for URL: data
PASS: Breakpoint should not be disabled initially.
Breakpoint disabled.
Sending XMLHttpRequest.
Wait for evaluate in page to return.
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: ShouldNotPauseOnXMLHttpRequestNotContainingURL
PASS: Added breakpoint for URL: nonexistant-url
PASS: Breakpoint should not be disabled initially.
Sending XMLHttpRequest.
Wait for evaluate in page to return.
PASS: Should not pause for breakpoint with different URL.
-- Running test teardown.

-- Running test case: RemoveBreakpoint
PASS: Added breakpoint for URL: data
PASS: Breakpoint should not be disabled initially.
Breakpoint removed.
Wait for evaluate in page to return.
PASS: Should not pause for removed breakpoint.
-- Running test teardown.

-- Running test case: RemoveDisabledBreakpoint
PASS: Added breakpoint for URL: data
PASS: Breakpoint should not be disabled initially.
Breakpoint disabled.
Breakpoint removed.
Wait for evaluate in page to return.
PASS: Should not pause for removed disabled breakpoint.
